refactor: insight cusotom repository module file refactoring
