+++ src/main/java/goorm/reinput/insight/repository/support/InsightRepositoryCustomImpl.java
@@ -1,7 +1,10 @@
-package goorm.reinput.insight.repository;
+package goorm.reinput.insight.repository.support;
 
+import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import goorm.reinput.insight.domain.Insight;
+import goorm.reinput.insight.domain.InsightRecommend;
+import goorm.reinput.insight.repository.CustomInsightRepository;
 import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,17 +19,30 @@ import static goorm.reinput.insight.domain.QInsight.insight;
 @Repository
 @Slf4j
 @RequiredArgsConstructor
-@Deprecated
-public class CustomInsightRepository {
+public class InsightRepositoryCustomImpl implements InsightRepositoryCustom{
     private final EntityManager em;
     private final JPAQueryFactory jpaQueryFactory;
 
     @Autowired
-    public CustomInsightRepository(EntityManager em) {
+    public InsightRepositoryCustomImpl(EntityManager em){
         this.em = em;
         this.jpaQueryFactory = new JPAQueryFactory(this.em);
     }
+    /*
+    * 5개의 랜덤 인사이트 반환
+     */
+    @Override
+    public Optional<List<Insight>> randInsight(){
 
+        List<Insight> insights = jpaQueryFactory.selectFrom(insight)
+                .orderBy(Expressions.numberTemplate(Double.class, "function('RAND')").asc())
+                .limit(5)
+                .fetch();
+
+        return Optional.ofNullable(insights);
+    }
+
+    @Override
     public Optional<List<Insight>> findByInsightFolderId(Long folderId) {
         log.info("[CustomInsightRepository] findByInsightFolderId {} called", folderId);
         return Optional.ofNullable(jpaQueryFactory
@@ -34,7 +50,4 @@ public class CustomInsightRepository {
                 .where(insight.folder.folderId.eq(folderId))
                 .fetch());
     }
-
-
-
 }
