+++ src/main/java/goorm/reinput/insight/service/InsightService.java
@@ -150,7 +150,7 @@ public class InsightService {
 
     public List<InsightSimpleResponseDto> getInsightListByFolderAndTag(Long userId, Long folderId, String tag) {
         // 폴더 ID로 Insight 리스트 조회
-        List<Insight> insightList = customInsightRepository.findByInsightFolderId(folderId).orElseGet(Collections::emptyList);
+        List<Insight> insightList = insightRepository.findByInsightFolderId(folderId).orElseGet(Collections::emptyList);
 
         // Insight 리스트에서 각 Insight의 hashTagList를 확인하여 주어진 태그를 부분적으로 포함하는 Insight만 필터링
         List<InsightSimpleResponseDto> filteredInsightList = insightList.stream()
@@ -253,7 +253,7 @@ public class InsightService {
 
     @Transactional
     public List<InsightSimpleResponseDto> getInsightList(Long userId, Long folderId) {
-        List<Insight> insightList = customInsightRepository.findByInsightFolderId(folderId).orElseGet(Collections::emptyList);
+        List<Insight> insightList = insightRepository.findByInsightFolderId(folderId).orElseGet(Collections::emptyList);
 
         return insightList.stream().map(insight -> {
